pyproject.toml
[project]
name = "audio-summary-uploader"
version = "0.1.0"
description = "NotebookLM → YouTube automation pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "pillow>=10.0",
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "loguru>=0.7",
    "notebooklm-py>=0.1",
    "playwright>=1.58.0",
]

[project.scripts]
automator = "automator.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/automator"]

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]
